requirements.txt
numpy==1.23.5
websocket-client==1.6.1
tensorflow==2.11.0
joblib==1.2.0
pylsl==1.16.0
keyboard==0.13.5